requirements.txt
slack_bolt
python-dotenv
requests
PyGithub
openai
